added stubgen
