+++ scripts/paths.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+PROJECT_DIRECTORY = Path(__file__).parent.parent
+DJANGO_SOURCE_DIRECTORY = PROJECT_DIRECTORY / "django-source"  # type: Path
